# documents/api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "GovTech Location Decision Maker API",
        "description": "GovTech Location Decision Maker APIs Information",
        "version": "1.0"
    },
    "servers": [{
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/v1/session/invite-users/{id}": {
            "get": {
                "tags": ["Session"],
                "summary": "Retrieve session invitees",
                "operationId": "getInvitees",
                "parameters": [{
                        "name": "id",
                        "in": "path",
                        "description": "Identifier of the session",
                        "required": true,
                        "schema": {
                            "title": "Session id",
                            "type": "string",
                            "description": "Identifier of the session"
                        }
                    }, {
                        "name": "Authorization",
                        "in": "header",
                        "description": "JWT Token : Bearer token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Bearer eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJ1c2Vyb25lIiwiaWF0IjoxNjk3MDc5MDkzLCJleHAiOjE2OTcxNjU0OTN9.RB70XduaPoPRDkF5HacJQyVr9GOFJZHMPM3gNJEWOY9nxMyNbie8iJ0N3ZCiN6Rnu-Wl-RsfyYgVh7xc3v_LXA"
                    }
                ],
                "responses": {
                    "406": {
                        "description": "Request not acceptable according to the Accept-* header.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Retrieve session invitees session invites information.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/SignInResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input supplied.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authorized. Application/User is not authorized to call this service.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Method other than POST is not supported.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error. Correct request but a problem occurred on the server. Dependent service is not available.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["Session"],
                "summary": "Send session invite to users",
                "operationId": "invite",
                "parameters": [{
                        "name": "id",
                        "in": "path",
                        "description": "Identifier of the session",
                        "required": true,
                        "schema": {
                            "title": "Session id",
                            "type": "string",
                            "description": "Identifier of the session"
                        }
                    }, {
                        "name": "Authorization",
                        "in": "header",
                        "description": "JWT Token : Bearer token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Bearer eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJ1c2Vyb25lIiwiaWF0IjoxNjk3MDc5MDkzLCJleHAiOjE2OTcxNjU0OTN9.RB70XduaPoPRDkF5HacJQyVr9GOFJZHMPM3gNJEWOY9nxMyNbie8iJ0N3ZCiN6Rnu-Wl-RsfyYgVh7xc3v_LXA"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSessionInviteRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "406": {
                        "description": "Request not acceptable according to the Accept-* header.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input supplied.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Return session invites information.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/SignInResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authorized. Application/User is not authorized to call this service.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Method other than POST is not supported.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error. Correct request but a problem occurred on the server. Dependent service is not available.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/account/{accountId}": {
            "get": {
                "tags": ["User"],
                "summary": "Retrieve user details of a user by account identifier",
                "operationId": "getByAccountId",
                "parameters": [{
                        "name": "accountId",
                        "in": "path",
                        "description": "Account identifier of the user",
                        "required": true,
                        "schema": {
                            "title": "User account id",
                            "type": "string",
                            "description": "Account identifier of the user"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return user information.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccountResponse"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "Request not acceptable according to the Accept-* header.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authorized. Application/User is not authorized to call this service.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error. Correct request but a problem occurred on the server. Dependent service is not available.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invalid account identifier supplied.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Method other than GET is not supported.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/account/{id}": {
            "get": {
                "tags": ["Account"],
                "summary": "Retrieve account details of a user by account identifier",
                "operationId": "getByAccountId_1",
                "parameters": [{
                        "name": "id",
                        "in": "path",
                        "description": "Account identifier of the user",
                        "required": true,
                        "schema": {
                            "title": "User account id",
                            "type": "string",
                            "description": "Account identifier of the user"
                        }
                    }
                ],
                "responses": {
                    "406": {
                        "description": "Request not acceptable according to the Accept-* header.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Return user account information.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccountResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authorized. Application/User is not authorized to call this service.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error. Correct request but a problem occurred on the server. Dependent service is not available.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invalid account identifier supplied.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Method other than GET is not supported.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "AccountResponse": {
                "title": "User account response",
                "type": "object",
                "properties": {
                    "id": {
                        "title": "User account identifier",
                        "type": "string",
                        "description": "Account identifier of the user"
                    },
                    "userName": {
                        "title": "User account name",
                        "type": "string",
                        "description": "Account name of the user"
                    }
                },
                "description": "Parameters for the user account response"
            },
            "ExceptionResponse": {
                "title": "Exception response",
                "type": "object",
                "properties": {
                    "message": {
                        "title": "Error message",
                        "type": "string",
                        "description": "Exception message"
                    },
                    "time": {
                        "title": "Exception occured time",
                        "type": "string",
                        "description": "Time at which the exceptoin has occured",
                        "format": "date-time"
                    }
                },
                "description": "Parameters for the exception response"
            },
...
            "SignInResponse": {
                "title": "User account sign in response",
                "type": "object",
                "properties": {
                    "id": {
                        "title": "User account id",
                        "type": "string",
                        "description": "Account identifier of the logged in user"
                    },
                    "accessToken": {
                        "title": "User account session token",
                        "type": "string",
                        "description": "Session access token of the logged in user"
                    }
                },
                "description": "Parameters for the user account sign in response"
            },
...
            "CreateSessionInviteRequest": {
                "required": ["userIds"],
                "type": "object",
                "properties": {
                    "userIds": {
                        "uniqueItems": true,
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
        }
    }
}